Mini Cart Validation and PDP add to cart function
