+++ GR/com/generic/tests/GR/HomePage/HPRegression.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="HomePageRegression" parallel="tests" thread-count="3" data-provider-thread-count="1">
+	<test name="HP Chrome">
+		<parameter name="browserName" value="Chrome"></parameter>
+		<parameter name="Env" value="prod"></parameter>
+		<parameter name="Brand" value="GR"></parameter>
+		<classes>
+			<class name="com.generic.tests.GR.HomePage.HomePageBase" />
+		</classes>
+	</test><!-- Test -->
+	<test name="HP iPad">
+		<parameter name="browserName" value="mobile_iPad"></parameter>
+		<parameter name="Env" value="prod"></parameter>
+		<parameter name="Brand" value="GR"></parameter>
+		<classes>
+			<class name="com.generic.tests.GR.HomePage.HomePageBase" />
+		</classes>
+	</test><!-- Test -->
+	<test name="HP iPhone">
+		<parameter name="browserName" value="mobile_iPhone X"></parameter>
+		<parameter name="Env" value="prod"></parameter>
+		<parameter name="Brand" value="GR"></parameter>
+		<classes>
+			<class name="com.generic.tests.GR.HomePage.HomePageBase" />
+		</classes>
+	</test><!-- Test -->
+</suite> <!-- Suite -->
