+++ GR/com/generic/tests/GR/HomePage/HPValidation.java
@@ -1,0 +1,18 @@
+package com.generic.tests.GR.HomePage;
+
+import com.generic.page.HomePage;
+import com.generic.setup.SelTestCase;
+
+public class HPValidation extends SelTestCase {
+
+	public static void logoValidate() throws Exception {
+		getCurrentFunctionName(true);
+		sassert().assertTrue(HomePage.validateLogodisplayed(), "Logo is not displayed");
+		HomePage.NavigateAwayFromHomePage(); 
+		HomePage.clickOnLogo(); 
+		sassert().assertTrue(HomePage.validateHomePageLink(), "Logo doesnt redirect to home page");
+		getCurrentFunctionName(false);
+		
+	}
+
+}
