+++ GR/com/generic/tests/GR/HomePage/HomePageBase.java
@@ -1,4 +1,4 @@
-package com.generic.tests.FG.HomePage;
+package com.generic.tests.GR.HomePage;
 
 import java.text.MessageFormat;
 import java.util.Arrays;
@@ -21,7 +21,6 @@ public class HomePageBase extends SelTestCase {
 
 	// possible scenarios
 	public static final String Logo = "Logo validation";
-	public static final String miniCart = "Mini cart validation";
 	public static final String verify = "verify";
 	public static final String header = "header";
 	public static final String footer = "footer";
@@ -64,12 +63,7 @@ public class HomePageBase extends SelTestCase {
 
 			if (proprties.contains(this.Logo)) {
 				LogoValidation.validate();
-			} else
-				if (proprties.contains(this.miniCart)) {
-					MiniCartValidation.validate();
-				}
-			
-					else{
+			} else {
 				Testlogs.get().debug("please check proprties provided in excel sheet");
 			}
