+++ src/com/generic/config/config.properties
@@ -24,7 +24,9 @@ toEmail=ibatta@pfsweb.com
 RandomItems=mat,shirt,tank,towel,tower,mirror,cake,short,fan,bed,shower,flip,flop,heart
 
 ############Pages URLs########################
-HomePage=https://www.frontgate.com
+HomePage=https://www.frontgate.com/
+searchURL=/ProductSearch2?searchTerm=test
+FG_PDP=/barrow-chesterfield-sofa/691936
 shoppingCartPage=/cart
 PLP=/bedding/basic-bedding-sets/c/19
 RegistrationPage=register 
@@ -47,3 +49,6 @@ signOutPage=https://hybrisdemo.conexus.co.uk:9002/yacceleratorstorefront/en/logo
 storeFinderPage=https://hybrisdemo.conexus.co.uk:9002/yacceleratorstorefront/en/store-finder
 errorPage=https://hybrisdemo.conexus.co.uk:9002/yacceleratorstorefront/en/j_spring_security_check
 CLP=https://hybrisdemo.conexus.co.uk:9002/yacceleratorstorefront/en/Collections/Streetwear/c/streetwear
+
+#######################################
+emptyCartText="shopping cart is empty"
