+++ src/com/generic/selector/HomePageSelectors.java
@@ -4,8 +4,18 @@ import com.generic.setup.cselector;
 
 public class HomePageSelectors
 {
-	public static final cselector logo = new cselector("css,#logo1 a.logo-anchor","css,#headerBox > div.u-flexbox.u-width-full.t-header-bar__title > div > a");
+	public static final cselector logo = new cselector("css,#logo1 a.logo-anchor","t-header__logo");
+	public static final String miniCartDeskt = "css,#cartAnchor";
+	public static final String miniCartMob = "css,.pw-header-bar__actions.t-header-bar__cart > button";
+	public static final cselector miniCartText = new cselector("empty-cart-label","css,.empty-cart-sign-in-container > .gwt-HTML.inst-copy","css,.m-mini-cart__empty-content.u-text-align-center>h2");
+	public static final String miniCartClose = "css,.pw-button__inner>#nav-icon";
+	public static final cselector miniCartProductContainer = new cselector("mini-cart-items-container","m-mini-cart__content-products");
+	public static final cselector miniCartCheckoutBtn = new cselector("css,.mini-cart-product-panel-btn","pw-link mini-cart__link pw-button pw--primary u-width-full u-text-weight-medium");		
+	//public static final cselector miniCartCheckoutBtn = new cselector("css,.mini-cart-product-panel-btn-img.checkout-version2","pw-link mini-cart__link pw-button pw--primary u-width-full u-text-weight-medium");		
+
+	//public static final cselector miniCartMobd = new cselector("css,#cartAnchor","css,.pw-header-bar__actions.t-header-bar__cart > button");
 	
+
 	//public static final String logo = "css,#logo1 a.logo-anchor";
 	
 	public static final String header = "header-container";
@@ -19,5 +29,4 @@ public class HomePageSelectors
 	public static final String footerLogo = "footer-ocm-logo";
 	public static final String footerEmailSignUp = "footer-email-sign-up";
 
-	
 }
