+++ src/com/generic/setup/GlobalVariables.java
@@ -14,6 +14,8 @@ public class GlobalVariables extends SelTestCase {
 		public static final String firefox = "Firefox";
 		public static final String chrome = "chrome";
 		public static final String Nexus = "mobile_Nexus 5";
+		public static final String iPad = "mobile_iPad";
+		public static final String iPhoneX = "mobile_iPhone X";
 		public static final String IE = "IE";
 		public static final String edge = "edge";
 		public static final String iOS = "IOS_grid";
